Add tables and mem-mapped I/O
This change adds a file with a table of all 8080 instructions and
attributes about them, from which we can generate tables for inclusion
in the C code of instruction mnemonics, sizes, and cycle counts. This
simplifies the core instruction execution loop and standardizes on the
logic in common for each instruction.

This change also adds rudimentary memory-mapped I/O. At an arbitrary
memory location, reads and writes are copied to standard input and
output, respectively. This is not an aspect of the 8080 itself, which
has its own IN and OUT instructions (which are yet to be implemented
here), but provides some utility during development and is useful for
exploring what an integrated computer using this emulated CPU might
require. The key abstraction here is adding a model of a bus, which
combines address and data buses in real computers, which handles the
actual I/O if the respective addresses are accessed, freeing the CPU
from having to track this (which it wouldn't in a von Neumann machine to
begin with).